make output of numbers and separators configurable
Being from germany, my google docs always complains / won't parse floats in the en-US format as we use the comma as decimal separator.
So I went out and made those thing configurable.
In addition one can set the column-separator now as well (ex. \t for tabstops or ; in case floats use comma like we do over here) and while I was at it: tune the float precision so we don't get 10 decimals if we don't want.

First I tried just loading the current culture information and not make it configurable but unity has some weird behaviour and nothing really returned de-DE for me (uiculture, installedculture, currentculture, etc...) so that's why it is a config föag now.

All defaults are unchanged as before: en-US, no rounding of numbers, comma as separator, but now there are options available.
Switched to List for the columns and also wanted to use String.Join before handing it over to the stringbuilder but somehow Unity refused to find System.String.Join so it's in a foreach with stringbuilder.Append() now.